improvments

+++ src/DataPacket.h
@@ -3,13 +3,14 @@
 
 #include <Arduino.h>
 
+// 32 bytes payload        10
 struct DataPacket {
-    uint8_t lx;
-    uint8_t ly;
-    bool    lb;
-    uint8_t rx;
-    uint8_t ry;
-    bool    rb;
+    uint16_t lx;        //  2
+    uint16_t ly;        //  2
+    bool     lb;        //  1
+    uint16_t rx;        //  2
+    uint16_t ry;        //  2
+    bool     rb;        //  1
 };
 
 #endif
